+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.9.3
+
+- Added manual crop options
+- Added basic saturation slider
+- Added more border options
+- Added original resolution export option
+- Added Input/Output .icc profile support
+- Added input icc profile for narrowband RGB (should mitigate common oversaturation issues)
+- Added horizontal & vertical flip options
+- Fix: respect file orientation.
+- UI redesign.
+
 ## 0.9.2
 
 - Make export consistent with preview (same demosaic + log bounds analysis)